web_interface/requirements.txt
Flask
psycopg2-binary
waitress
toml
requests
